pyproject.toml
[project]
name = "clt"
description = "Crono's collection of tools"
version = "2025.07.16"
authors = [{ name = "Travis Kern", email = "traviskern@adt.com" }]
maintainers = [{ name = "Travis Kern", email = "crono@cronolabs.net" }]
readme = "README.md"
requires-python = ">= 3.12"

dependencies = [
	"ruff",
	"pyright",
	"httpx",
	"pre-commit",
	"tabulate>=0.9.0",
	"pyyaml",
	"icmplib>=3.0.4",
	"tqdm>=4.67.1",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
